Adding RBAC items

+++ controllers/ItemsController.php
@@ -18,6 +18,7 @@ use cinghie\articles\models\ItemsSearch;
 use yii\filters\AccessControl;
 use yii\filters\VerbFilter;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 
 /**
@@ -71,9 +72,15 @@ class ItemsController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        // Check RBAC Permission
+        if(Yii::$app->user->can('view-articles'))
+        {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -83,57 +90,63 @@ class ItemsController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Items();
+        // Check RBAC Permission
+        if(Yii::$app->user->can('create-articles'))
+        {
+            $model = new Items();
 
-        if ( $model->load(Yii::$app->request->post()) ) {
-			
-			// Set Modified as actual date 
-			$model->modified = date("Y-m-d H:i:s");
-			
-			// If alias is not set, generate it
-			if ($_POST['Items']['alias']=="") 
-			{
-				$model->alias = $model->generateAlias($model->title);
-			}
-			
-			// Upload Image and Thumb if is not Null
-			$imagePath   = Yii::getAlias(Yii::$app->controller->module->itemImagePath);
-			$thumbPath   = Yii::getAlias(Yii::$app->controller->module->itemThumbPath);
-			$imgNameType = Yii::$app->controller->module->imageNameType;
-			$imgOptions  = Yii::$app->controller->module->thumbOptions;
-			$imgName     = $model->title;
-			$fileField   = "image";
-			
-			// Create UploadFile Instance
-			$image = $model->uploadFile($imgName,$imgNameType,$imagePath,$fileField);
-			
-			if ($model->save()) {	
-				
-				// upload only if valid uploaded file instance found
-                if ($image !== false) 
-				{
-					// save thumbs to thumbPaths
-					$thumb = $model->createThumbImages($image,$imagePath,$imgOptions,$thumbPath);
+            if ( $model->load(Yii::$app->request->post()) )
+            {
+                // Set Modified as actual date
+                $model->modified = date("Y-m-d H:i:s");
+
+                // If alias is not set, generate it
+                if ($_POST['Items']['alias']=="")
+                {
+                    $model->alias = $model->generateAlias($model->title);
                 }
-				
-				// Set Success Message
-				Yii::$app->session->setFlash('success', Yii::t('articles', 'Item has been created!'));
-				
-				return $this->redirect(['index']);
-			
-			} else {
-				
-				// Set Error Message
-				Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
-				
-				return $this->render('create', ['model' => $model,]);
+
+                // Upload Image and Thumb if is not Null
+                $imagePath   = Yii::getAlias(Yii::$app->controller->module->itemImagePath);
+                $thumbPath   = Yii::getAlias(Yii::$app->controller->module->itemThumbPath);
+                $imgNameType = Yii::$app->controller->module->imageNameType;
+                $imgOptions  = Yii::$app->controller->module->thumbOptions;
+                $imgName     = $model->title;
+                $fileField   = "image";
+
+                // Create UploadFile Instance
+                $image = $model->uploadFile($imgName,$imgNameType,$imagePath,$fileField);
+
+                if ($model->save()) {
+
+                    // upload only if valid uploaded file instance found
+                    if ($image !== false)
+                    {
+                        // save thumbs to thumbPaths
+                        $thumb = $model->createThumbImages($image,$imagePath,$imgOptions,$thumbPath);
+                    }
+
+                    // Set Success Message
+                    Yii::$app->session->setFlash('success', Yii::t('articles', 'Item has been created!'));
+
+                    return $this->redirect(['index']);
+
+                } else {
+
+                    // Set Error Message
+                    Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
+
+                    return $this->render('create', ['model' => $model,]);
+                }
+
+            } else {
+
+                Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
+
+                return $this->render('create', ['model' => $model,]);
             }
-		
-		} else {
-			
-			Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
-            
-			return $this->render('create', ['model' => $model,]);
+        } else {
+            throw new ForbiddenHttpException;
         }
 			
     }
@@ -146,57 +159,61 @@ class ItemsController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        // Check RBAC Permission
+        if(Yii::$app->user->can('update-articles'))
+        {
+            $model = $this->findModel($id);
 
-        if ( $model->load(Yii::$app->request->post()) ) {
-			
-			// Set Modified as actual date 
-			$model->modified = date("Y-m-d H:i:s");
-			
-			// If alias is not set, generate it
-			if ($_POST['Items']['alias']=="") 
-			{
-				$model->alias = $model->generateAlias($model->title);
-			}
-			
-			// Upload Image and Thumb if is not Null
-			$imagePath   = Yii::getAlias(Yii::$app->controller->module->itemImagePath);
-			$thumbPath   = Yii::getAlias(Yii::$app->controller->module->itemThumbPath);
-			$imgNameType = Yii::$app->controller->module->imageNameType;
-			$imgOptions  = Yii::$app->controller->module->thumbOptions;
-			$imgName     = $model->title;
-			$fileField   = "image";
-			
-			// Create UploadFile Instance
-			$image = $model->uploadFile($imgName,$imgNameType,$imagePath,$fileField);	
-						
-			if ($model->save()) {			
-				
-				// upload only if valid uploaded file instance found
-                if ($image !== false) 
-				{
-					// save thumbs to thumbPaths
-					$thumb = $model->createThumbImages($image,$imagePath,$imgOptions,$thumbPath);
+            if ($model->load(Yii::$app->request->post())) {
+
+                // Set Modified as actual date
+                $model->modified = date("Y-m-d H:i:s");
+
+                // If alias is not set, generate it
+                if ($_POST['Items']['alias'] == "") {
+                    $model->alias = $model->generateAlias($model->title);
                 }
-				
-				// Set Success Message
-				Yii::$app->session->setFlash('success', Yii::t('articles', 'Item has been updated!'));
-				
-				return $this->redirect(['index']);
-			
-			} else {
-				
-				// Set Error Message
-				Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
-				
-				return $this->render('update', ['model' => $model,]);
+
+                // Upload Image and Thumb if is not Null
+                $imagePath = Yii::getAlias(Yii::$app->controller->module->itemImagePath);
+                $thumbPath = Yii::getAlias(Yii::$app->controller->module->itemThumbPath);
+                $imgNameType = Yii::$app->controller->module->imageNameType;
+                $imgOptions = Yii::$app->controller->module->thumbOptions;
+                $imgName = $model->title;
+                $fileField = "image";
+
+                // Create UploadFile Instance
+                $image = $model->uploadFile($imgName, $imgNameType, $imagePath, $fileField);
+
+                if ($model->save()) {
+
+                    // upload only if valid uploaded file instance found
+                    if ($image !== false) {
+                        // save thumbs to thumbPaths
+                        $thumb = $model->createThumbImages($image, $imagePath, $imgOptions, $thumbPath);
+                    }
+
+                    // Set Success Message
+                    Yii::$app->session->setFlash('success', Yii::t('articles', 'Item has been updated!'));
+
+                    return $this->redirect(['index']);
+
+                } else {
+
+                    // Set Error Message
+                    Yii::$app->session->setFlash('error', Yii::t('articles', 'Item could not be saved!'));
+
+                    return $this->render('update', ['model' => $model,]);
+                }
+
+            } else {
+
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
             }
-			
         } else {
-			
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            throw new ForbiddenHttpException;
         }
     }
 
@@ -208,20 +225,25 @@ class ItemsController extends Controller
      */
     public function actionDelete($id)
     {
-        $model = $this->findModel($id);
-		
-		if ($model->delete()) 
-		{
-            if (!$model->deleteImage()) {
-                Yii::$app->session->setFlash('error', 'Error deleting image');
-            } else {
-				Yii::$app->session->setFlash('success', Yii::t('articles', 'Category has been deleted!'));
-			}
-        } else {
-			Yii::$app->session->setFlash('error', 'Error deleting item');
-		}
+        // Check RBAC Permission
+        if(Yii::$app->user->can('delete-articles'))
+        {
+            $model = $this->findModel($id);
 
-        return $this->redirect(['index']);
+            if ($model->delete()) {
+                if (!$model->deleteImage()) {
+                    Yii::$app->session->setFlash('error', 'Error deleting image');
+                } else {
+                    Yii::$app->session->setFlash('success', Yii::t('articles', 'Category has been deleted!'));
+                }
+            } else {
+                Yii::$app->session->setFlash('error', 'Error deleting item');
+            }
+
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 	
 	/**
@@ -232,20 +254,25 @@ class ItemsController extends Controller
      */
 	public function actionDeleteimage($id) 
 	{
-		$model = $this->findModel($id);
-		
-		if ($model->deleteImage()) 
-		{
-			$model->image = "";
-			$model->save();
-			Yii::$app->session->setFlash('success', Yii::t('articles', 'The image was removed successfully! Now, you can upload another by clicking Browse in the Image Tab.'));
-		} else {
-			Yii::$app->session->setFlash('error', Yii::t('articles', 'Error removing image. Please try again later or contact the system admin.'));
-		}
-		
-		return $this->redirect([
-				'update', 'id' => $model->id,
-		]);
+        // Check RBAC Permission
+        if(Yii::$app->user->can('update-articles'))
+        {
+            $model = $this->findModel($id);
+
+            if ($model->deleteImage()) {
+                $model->image = "";
+                $model->save();
+                Yii::$app->session->setFlash('success', Yii::t('articles', 'The image was removed successfully! Now, you can upload another by clicking Browse in the Image Tab.'));
+            } else {
+                Yii::$app->session->setFlash('error', Yii::t('articles', 'Error removing image. Please try again later or contact the system admin.'));
+            }
+
+            return $this->redirect([
+                'update', 'id' => $model->id,
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
+        }
 	}
 
     /**
